dhcp6: moved most of the sources to src/dhcp6
