+++ dhcp6/dbus-api.h
@@ -23,7 +23,7 @@
 #define __WICKED_DHCP6_DBUS_API_H__
 
 #include <wicked/dbus.h>
-#include <dhcp6/dhcp6.h>
+#include "dhcp6/dhcp.h"
 
 extern void				ni_objectmodel_dhcp6_init(void);
 extern ni_dbus_object_t *		ni_objectmodel_register_dhcp6_device(ni_dbus_server_t *, ni_dhcp6_device_t *);
