+++ src/Makefile.am
@@ -154,8 +154,12 @@ libwicked_dhcp4_la_SOURCES		= \
 
 libwicked_dhcp6_la_CFLAGS		= $(libwicked_la_CFLAGS)
 libwicked_dhcp6_la_SOURCES		= \
+	dhcp6/device.c		\
+	dhcp6/fsm.c		\
 	dhcp6/lease.c		\
-	dhcp6/options.c
+	dhcp6/options.c		\
+	dhcp6/protocol.c	\
+	dhcp6/tester.c
 
 libwicked_client_la_CFLAGS		= $(libwicked_la_CFLAGS)
 libwicked_client_la_SOURCES		= \
@@ -189,8 +193,13 @@ wicked_headers			= \
 	dhcp4/lease.h		\
 	dhcp4/protocol.h	\
 	dhcp4/tester.h		\
+	dhcp6/device.h		\
+	dhcp6/dhcp.h		\
 	dhcp6/lease.h		\
+	dhcp6/fsm.h		\
 	dhcp6/options.h		\
+	dhcp6/protocol.h	\
+	dhcp6/tester.h		\
 	duid.h			\
 	ibft.h			\
 	ipv6_priv.h		\
