+++ src/dhcp6/fsm.c
@@ -30,7 +30,7 @@
 #include <wicked/logging.h>
 #include <wicked/resolver.h>
 
-#include "dhcp6/dhcp6.h"
+#include "dhcp6/dhcp.h"
 #include "dhcp6/device.h"
 #include "dhcp6/protocol.h"
 #include "dhcp6/fsm.h"
