+++ src/dhcp6/protocol.c
@@ -46,7 +46,7 @@
 #include <wicked/xml.h>
 #endif
 
-#include "dhcp6/dhcp6.h"
+#include "dhcp6/dhcp.h"
 #include "dhcp6/device.h"
 #include "dhcp6/protocol.h"
 #include "dhcp6/fsm.h"
